[MOV] survey: prepare file for refactoring
Move the survey_result_pagination public widget from /src/js/survey_result.js
to /src/interactions/survey_result_pagination.js.
Extract the code for the entries of the paginated result table in a separated
template (will be used by the new interaction).
This commit is added to make it easier to follow the history.
task-4367641
