+++ addons/survey/__manifest__.py
@@ -71,7 +71,6 @@ sent mails with personal token for the invitation of the survey.
             'survey/static/src/js/survey_form.js',
             'survey/static/src/js/survey_preload_image_mixin.js',
             'survey/static/src/js/survey_result.js',
-            'survey/static/src/interactions/survey_result_chart.js',
             ('include', 'web._assets_helpers'),
             ('include', 'web._assets_frontend_helpers'),
             'web/static/src/scss/pre_variables.scss',
