+++ addons/survey/static/src/interactions/survey_result_pagination.js
@@ -1,0 +1,80 @@
+import publicWidget from "@web/legacy/js/public/public_widget";
+
+// TODO awa: this widget loads all records and only hides some based on page
+// -> this is ugly / not efficient, needs to be refactored
+publicWidget.registry.SurveyResultPagination = publicWidget.Widget.extend({
+    events: {
+        'click li.o_survey_js_results_pagination a': '_onPageClick',
+        "click .o_survey_question_answers_show_btn": "_onShowAllAnswers",
+    },
+
+    //--------------------------------------------------------------------------
+    // Widget
+    //--------------------------------------------------------------------------
+
+    /**
+     * @override
+     * @param {$.Element} params.questionsEl The element containing the actual questions
+     *   to be able to hide / show them based on the page number
+     */
+    init: function (parent, params) {
+        this._super.apply(this, arguments);
+        this.$questionsEl = params.questionsEl;
+    },
+
+    /**
+     * @override
+     */
+    start: function () {
+        var self = this;
+        return this._super.apply(this, arguments).then(function () {
+            self.limit = self.$el.data("record_limit");
+        });
+    },
+
+    // -------------------------------------------------------------------------
+    // Handlers
+    // -------------------------------------------------------------------------
+
+    /**
+     * @private
+     * @param {MouseEvent} ev
+     */
+    _onPageClick: function (ev) {
+        ev.preventDefault();
+        this.$('li.o_survey_js_results_pagination').removeClass('active');
+
+        var $target = $(ev.currentTarget);
+        $target.closest('li').addClass('active');
+        this.$questionsEl.find("tbody tr").addClass("d-none");
+
+        var num = $target.text();
+        var min = this.limit * (num - 1) - 1;
+        if (min === -1) {
+            this.$questionsEl
+                .find("tbody tr:lt(" + this.limit * num + ")")
+                .removeClass("d-none");
+        } else {
+            this.$questionsEl
+                .find("tbody tr:lt(" + this.limit * num + "):gt(" + min + ")")
+                .removeClass("d-none");
+        }
+    },
+
+    /**
+     * @private
+     * @param {MouseEvent} ev
+     */
+    _onShowAllAnswers: function (ev) {
+        const btnEl = ev.currentTarget;
+        const pager = btnEl.previousElementSibling;
+        btnEl.classList.add("d-none");
+        this.$questionsEl.find("tbody tr").removeClass("d-none");
+        pager.classList.add("d-none");
+        this.$questionsEl.parent().addClass("h-auto");
+    },
+});
+
+export default {
+    paginationWidget: publicWidget.registry.SurveyResultPagination,
+};
